.claude/envoy/requirements.txt
requests>=2.28.0
google-genai>=0.8.0
